# src/main/java/atlassports/api/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Users", description = "User profile operations")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_USERS_ME_GET = "/users/me";

    /**
     * GET /users/me : Get current user profile
     *
     * @return User profile (status code 200)
     */
    @Operation(
            operationId = "usersMeGet",
            summary = "Get current user profile",
            tags = {"Users"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "User profile", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = UsersApi.PATH_USERS_ME_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Object> usersMeGet(

    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_ME_PATCH = "/users/me";

    /**
     * PATCH /users/me : Update current user profile
     *
     * @param body (required)
     * @return Updated user (status code 200)
     */
    @Operation(
            operationId = "usersMePatch",
            summary = "Update current user profile",
            tags = {"Users"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Updated user", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PATCH,
            value = UsersApi.PATH_USERS_ME_PATCH,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<Object> usersMePatch(
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}