# src/main/java/atlassports/api/VenuesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import atlassports.model.openapi.VenuesGet200Response;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.lang.Nullable;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.math.BigDecimal;
import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Venues", description = "Venue management")
public interface VenuesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_VENUES_GET = "/venues";

    /**
     * GET /venues : Get all venues
     *
     * @param page       (optional, default to 0)
     * @param size       (optional, default to 20)
     * @param sort       (optional)
     * @param city       (optional)
     * @param type       (optional)
     * @param minPrice   (optional)
     * @param maxPrice   (optional)
     * @param hasParking (optional)
     * @return List of venues (status code 200)
     */
    @Operation(
            operationId = "venuesGet",
            summary = "Get all venues",
            tags = {"Venues"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of venues", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = VenuesGet200Response.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = VenuesApi.PATH_VENUES_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<VenuesGet200Response> venuesGet(
            @Min(value = 0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
            @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
            @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) @Nullable String sort,
            @Parameter(name = "city", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "city", required = false) @Nullable String city,
            @Parameter(name = "type", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "type", required = false) @Nullable String type,
            @Parameter(name = "minPrice", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "minPrice", required = false) @Nullable BigDecimal minPrice,
            @Parameter(name = "maxPrice", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "maxPrice", required = false) @Nullable BigDecimal maxPrice,
            @Parameter(name = "hasParking", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "hasParking", required = false) @Nullable Boolean hasParking
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 6, \"last\" : true, \"totalPages\" : 5, \"page\" : 0, \"content\" : [ null, null ], \"first\" : true, \"totalElements\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_VENUES_POST = "/venues";

    /**
     * POST /venues : Create a venue
     *
     * @param body (required)
     * @return Venue created (status code 201)
     */
    @Operation(
            operationId = "venuesPost",
            summary = "Create a venue",
            tags = {"Venues"},
            responses = {
                    @ApiResponse(responseCode = "201", description = "Venue created", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = VenuesApi.PATH_VENUES_POST,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<Object> venuesPost(
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_VENUES_VENUE_ID_GET = "/venues/{venueId}";

    /**
     * GET /venues/{venueId} : Get venue details
     *
     * @param venueId (required)
     * @return Venue details (status code 200)
     */
    @Operation(
            operationId = "venuesVenueIdGet",
            summary = "Get venue details",
            tags = {"Venues"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Venue details", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = VenuesApi.PATH_VENUES_VENUE_ID_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Object> venuesVenueIdGet(
            @NotNull @Parameter(name = "venueId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("venueId") UUID venueId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_VENUES_VENUE_ID_PATCH = "/venues/{venueId}";

    /**
     * PATCH /venues/{venueId} : Update venue
     *
     * @param venueId (required)
     * @param body    (required)
     * @return Updated venue (status code 200)
     */
    @Operation(
            operationId = "venuesVenueIdPatch",
            summary = "Update venue",
            tags = {"Venues"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Updated venue", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.PATCH,
            value = VenuesApi.PATH_VENUES_VENUE_ID_PATCH,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<Object> venuesVenueIdPatch(
            @NotNull @Parameter(name = "venueId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("venueId") UUID venueId,
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}