# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Atlas Sports API
  version: 1.0.0
servers:
- url: http://localhost:8080/api
...
paths:
...
  /users/me:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
          description: User profile
      summary: Get current user profile
      tags:
      - Users
      x-accepts:
      - application/json
      x-tags:
      - tag: Users
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertUserDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
          description: Updated user
      summary: Update current user profile
      tags:
      - Users
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Users
...
  /venues:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      - explode: true
        in: query
        name: city
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: type
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: minPrice
        required: false
        schema:
          type: number
        style: form
      - explode: true
        in: query
        name: maxPrice
        required: false
        schema:
          type: number
        style: form
      - explode: true
        in: query
        name: hasParking
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_venues_get_200_response"
          description: List of venues
      summary: Get all venues
      tags:
      - Venues
      x-accepts:
      - application/json
      x-tags:
      - tag: Venues
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVenueDto"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDto"
          description: Venue created
      summary: Create a venue
      tags:
      - Venues
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Venues
  /venues/{venueId}:
    get:
      parameters:
      - explode: false
        in: path
        name: venueId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDto"
          description: Venue details
      summary: Get venue details
      tags:
      - Venues
      x-accepts:
      - application/json
      x-tags:
      - tag: Venues
    patch:
      parameters:
      - explode: false
        in: path
        name: venueId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVenueDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDto"
          description: Updated venue
      summary: Update venue
      tags:
      - Venues
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Venues
...
components:
...
  schemas:
...
    UserDto:
      type: object
    UpsertUserDto:
      type: object
...
    VenueDto:
      type: object
    UpsertVenueDto:
      type: object
...
    PageResponse:
      properties:
        content:
          items:
            type: object
          type: array
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
      type: object
...
    _venues_get_200_response:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/VenueDto"
            type: array
        type: object
      example:
        size: 6
        last: true
        totalPages: 5
        page: 0
        content:
        - null
        - null
        first: true
        totalElements: 1
...